Fixing Error

+++ lib/model/weather_forecast_model.dart
@@ -142,7 +142,7 @@ class Lista {
         weather!.add(new Weather.fromJson(v));
       });
     }
-    speed = json['speed'];
+    speed = json['speed'].toDouble();
     deg = json['deg'];
     gust = json['gust'].toDouble();
     clouds = json['clouds'];
